Do not check for SSL with secure option
The W3 Secure Contexts specification defines additional
contexts to be potentially secure beyond HTTPS, such as
localhost, 127.0.0.1, and *.localhost. Therefore, by
limiting secure cookies to only SSL, we are restricting
session uses in additional valid contexts (which is
particularly important because setting SameSite=None,
which is required by cross-domain iframes, is only doable
with the Secure flag).

Backwards-compatibility wise this should be safe because
Secure cookies were never returned outside of SSL, and now
we will start to. Furthermore, if a Secure cookie is returned
over a non-Secure Context, browser automatically discard them.

https://developer.mozilla.org/en-US/docs/Web/Security/Secure_Contexts